Added automatic palette generation to conv_to_sv
So I figured that palette generation was a little too... manual, and took it upon myself to automagically generate a minimal palette from the colors in the image. It works by basically lumping colors within a certain distance of one another together, thus saving room in the palette. Then, I rolled this palette into the SV generation code to make it super easy to use. Also added prompts to show the usage of the script, etc. Try it with GalagaLogo.png, an image I'm PR'ing into sprite_originals